+++ README.md
@@ -2,5 +2,5 @@
 
 ## 1단계 - 캐시 적용
 ### 기능 요구 사항
-- [ ] 지하철 노선도 조회 시 ETag를 통해 캐시를 적용
-- [ ] LineControllerTest의 ETag 테스트를 성공 시키기
+- [x] 지하철 노선도 조회 시 ETag를 통해 캐시를 적용
+- [x] LineControllerTest의 ETag 테스트를 성공 시키기
