--- src/main/java/wooteco/subway/admin/config/ETagHeaderFilter.java
@@ -1,5 +1,0 @@
-package wooteco.subway.admin.config;
-
-// TODO: ETag 관련 설정하기
-public class ETagHeaderFilter {
-}
