+++ src/main/java/wooteco/subway/admin/controller/LineController.java
@@ -1,17 +1,24 @@
 package wooteco.subway.admin.controller;
 
+import java.net.URI;
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineRequest;
 import wooteco.subway.admin.dto.LineResponse;
 import wooteco.subway.admin.dto.LineStationCreateRequest;
-import wooteco.subway.admin.dto.WholeSubwayResponse;
 import wooteco.subway.admin.service.LineService;
-
-import java.net.URI;
-import java.util.List;
 
 @RequestMapping("/lines")
 @RestController
@@ -27,19 +34,14 @@ public class LineController {
         Line persistLine = lineService.save(view.toLine());
 
         return ResponseEntity
-                .created(URI.create("/lines/" + persistLine.getId()))
-                .body(LineResponse.of(persistLine));
+            .created(URI.create("/lines/" + persistLine.getId()))
+            .body(LineResponse.of(persistLine));
     }
 
     @GetMapping
     public ResponseEntity<List<LineResponse>> showLine() {
         lineService.showLines();
         return ResponseEntity.ok().body(LineResponse.listOf(lineService.showLines()));
-    }
-
-    @GetMapping("/detail")
-    public ResponseEntity<WholeSubwayResponse> retrieveLines() {
-        return ResponseEntity.ok().body(lineService.wholeLines());
     }
 
     @GetMapping("/{id}")
@@ -60,13 +62,15 @@ public class LineController {
     }
 
     @PostMapping("/{lineId}/stations")
-    public ResponseEntity addLineStation(@PathVariable Long lineId, @RequestBody LineStationCreateRequest view) {
+    public ResponseEntity addLineStation(@PathVariable Long lineId,
+        @RequestBody LineStationCreateRequest view) {
         lineService.addLineStation(lineId, view);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{lineId}/stations/{stationId}")
-    public ResponseEntity removeLineStation(@PathVariable Long lineId, @PathVariable Long stationId) {
+    public ResponseEntity removeLineStation(@PathVariable Long lineId,
+        @PathVariable Long stationId) {
         lineService.removeLineStation(lineId, stationId);
         return ResponseEntity.noContent().build();
     }
