--- src/test/java/wooteco/subway/admin/SubwayApplicationTests.java
@@ -1,13 +1,0 @@
-package wooteco.subway.admin;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SubwayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
